N64 work
-. Added mempak for N64 (seems functional)
-. Fixed deadzone for N64
-. Left trigger thing...

N64 rumble can now be toggled for a mempak with the BACK button. When
using the mempak, the Scroll Lock LED will be on.

Following nRage, mempak files will be saved on the MemPaks directory.
It's a raw dump with the mempak extention (I think it might work as
.mpk)

Also, now if you hold the Left Trigger, the face buttons become
C-Buttons. Useful in case you need to press opposite buttons
simultaneously.

And... Y/B buttons are now A and B buttons too, respectively.